moved player to player folder
