--- core/player.go
@@ -1,11 +1,0 @@
-package core
-
-type Player struct {
-	Deck *Deck
-}
-
-func NewPlayer() *Player {
-	return &Player{
-		Deck: NewDeck(),
-	}
-}
